Move contents of pycds/__init__.py to other modules
